watch: split implementation to aseparate file
Split CDBA's watching code to a separate file.
