+++ cdba-server.c
@@ -4,8 +4,6 @@
  *
  * SPDX-License-Identifier: BSD-3-Clause
  */
-#include <sys/time.h>
-#include <alloca.h>
 #include <err.h>
 #include <errno.h>
 #include <fcntl.h>
@@ -22,8 +20,8 @@
 #include "device_parser.h"
 #include "fastboot.h"
 #include "list.h"
+#include "watch.h"
 
-static bool quit_invoked;
 static const char *username;
 
 struct device *selected_device;
@@ -88,7 +86,7 @@ static void msg_select_board(const void *param)
 	selected_device = device_open(param, username, &fastboot_ops);
 	if (!selected_device) {
 		fprintf(stderr, "failed to open %s\n", (const char *)param);
-		quit_invoked = true;
+		watch_quit();
 	}
 
 	cdba_send(MSG_SELECT_BOARD);
@@ -224,119 +222,14 @@ static int handle_stdin(int fd, void *buf)
 	return 0;
 }
 
-struct watch {
-	struct list_head node;
-
-	int fd;
-	int (*cb)(int, void*);
-	void *data;
-};
-
-struct timer {
-	struct list_head node;
-	struct timeval tv;
-
-	void (*cb)(void *);
-	void *data;
-};
-
-static struct list_head read_watches = LIST_INIT(read_watches);
-static struct list_head timer_watches = LIST_INIT(timer_watches);
-
-void watch_add_readfd(int fd, int (*cb)(int, void*), void *data)
-{
-	struct watch *w;
-
-	w = calloc(1, sizeof(*w));
-	w->fd = fd;
-	w->cb = cb;
-	w->data = data;
-
-	list_add(&read_watches, &w->node);
-}
-
-void watch_timer_add(int timeout_ms, void (*cb)(void *), void *data)
-{
-	struct timeval tv_timeout;
-	struct timeval now;
-	struct timer *t;
-
-	t = calloc(1, sizeof(*t));
-
-	gettimeofday(&now, NULL);
-
-	tv_timeout.tv_sec = timeout_ms / 1000;
-	tv_timeout.tv_usec = (timeout_ms % 1000) * 1000;
-
-	t->cb = cb;
-	t->data = data;
-	timeradd(&now, &tv_timeout, &t->tv);
-
-	list_add(&timer_watches, &t->node);
-}
-
-static struct timeval *watch_timer_next(void)
-{
-	static struct timeval timeout;
-	struct timeval now;
-	struct timer *next;
-	struct timer *t;
-
-	if (list_empty(&timer_watches))
-		return NULL;
-
-	next = list_entry_first(&timer_watches, struct timer, node);
-
-	list_for_each_entry(t, &timer_watches, node) {
-		if (timercmp(&t->tv, &next->tv, <))
-			next = t;
-	}
-
-	gettimeofday(&now, NULL);
-	timersub(&next->tv, &now, &timeout);
-	if (timeout.tv_sec < 0) {
-		timeout.tv_sec = 0;
-		timeout.tv_usec = 0;
-	}
-
-	return &timeout;
-}
-
-static void watch_timer_invoke(void)
-{
-	struct timeval now;
-	struct timer *tmp;
-	struct timer *t;
-
-	gettimeofday(&now, NULL);
-
-	list_for_each_entry_safe(t, tmp, &timer_watches, node) {
-		if (timercmp(&t->tv, &now, <)) {
-			t->cb(t->data);
-
-			list_del(&t->node);
-			free(t);
-		}
-	}
-}
-
 static void sigpipe_handler(int signo)
 {
-	quit_invoked = true;
-}
-
-void watch_quit(void)
-{
-	quit_invoked = true;
+	watch_quit();
 }
 
 int main(int argc, char **argv)
 {
-	struct timeval *timeoutp;
-	struct watch *w;
-	fd_set rfds;
 	int flags;
-	int nfds;
 	int ret;
 
 	signal(SIGPIPE, sigpipe_handler);
@@ -365,40 +258,7 @@ int main(int argc, char **argv)
 	flags = fcntl(STDIN_FILENO, F_GETFL, 0);
 	fcntl(STDIN_FILENO, F_SETFL, flags | O_NONBLOCK);
 
-	while (!quit_invoked) {
-		nfds = 0;
-
-		list_for_each_entry(w, &read_watches, node) {
-			nfds = MAX(nfds, w->fd);
-			FD_SET(w->fd, &rfds);
-		}
-
-		if (!FD_ISSET(STDIN_FILENO, &rfds)) {
-			fprintf(stderr, "rfds is trash!\n");
-			goto done;
-		}
-
-		timeoutp = watch_timer_next();
-		ret = select(nfds + 1, &rfds, NULL, NULL, timeoutp);
-		if (ret < 0 && errno == EINTR)
-			continue;
-		else if (ret < 0)
-			break;
-
-		watch_timer_invoke();
-
-		list_for_each_entry(w, &read_watches, node) {
-			if (FD_ISSET(w->fd, &rfds)) {
-				ret = w->cb(w->fd, w->data);
-				if (ret < 0) {
-					fprintf(stderr, "cb returned %d\n", ret);
-					goto done;
-				}
-			}
-		}
-	}
-
-done:
+	watch_run();
 
 	/* if we got here, stdin/out/err might be not accessible anymore */
 	ret = open("/dev/null", O_RDWR);
