+++ device.c
@@ -23,6 +23,7 @@
 #include "list.h"
 #include "ppps.h"
 #include "status-cmd.h"
+#include "watch.h"
 
 #define ARRAY_SIZE(x) ((sizeof(x)/sizeof((x)[0])))
