+++ fastboot.c
@@ -22,6 +22,7 @@
 
 #include "cdba-server.h"
 #include "fastboot.h"
+#include "watch.h"
 
 #define MAX_USBFS_BULK_SIZE (16*1024)
